requirements.txt
twitchio
keyboard
ollama
soundfile
pygame
mutagen
pydub
gtts
discord.py[voice]
python-dotenv
